rag/requirements.txt
tiktoken
loguru
langchain
streamlit
faiss-cpu
huggingface-hub
langchain-community
pandas
sentence-transformers
openai
fastapi
uvicorn
python-dotenv